Backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
backoff==2.2.1
bcrypt==4.2.0
boto3==1.20.26
botocore
build==1.2.2.post1
cachetools==4.2.4
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.16
chromadb-client==0.5.16
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==41.0.7
Deprecated==1.2.14
dj-database-url==2.1.0
Django==5.1.2
django-anymail==10.2
django-ckeditor-5==0.2.15
django-cors-headers==3.14.0
django-googledrive-storage==1.6.0
django-jazzmin==2.6.0
django-storages==1.12.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
durationpy==0.9
environs==10.0.0
exceptiongroup==1.2.2
fastapi==0.115.4
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
google-api-core==2.10.2
google-api-python-client==2.149.0
google-auth==1.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.65.0
grpcio==1.67.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
inflection==0.5.1
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
kubernetes==31.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.10
overrides==7.7.0
packaging==24.1
pillow==11.0.0
posthog==3.7.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
PyJWT==2.6.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.3
rsa==4.9
s3transfer==0.5.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
shellingham==1.5.4
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.41.2
stripe==7.9.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.1
tomli==2.0.2
torch==2.5.1
tqdm==4.66.6
transformers==4.46.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3
uvicorn==0.32.0
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.1
wrapt==1.16.0
zipp==3.20.2
